Refactored to modularize the code
